AB#2028938 pop up a message when run query with null workspace. show reminder in table.emptyText when workspace is null
