+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-monitor/src/main/java/com/microsoft/azure/toolkit/intellij/monitor/view/right/MonitorLogTablePanel.java
@@ -15,7 +15,6 @@ import com.intellij.openapi.fileChooser.FileChooserFactory;
 import com.intellij.openapi.fileChooser.FileSaverDescriptor;
 import com.intellij.openapi.fileChooser.FileSaverDialog;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.project.ProjectManager;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.openapi.vfs.VirtualFileWrapper;
@@ -25,13 +24,11 @@ import com.intellij.ui.components.AnActionLink;
 import com.intellij.util.ui.JBUI;
 import com.microsoft.azure.toolkit.intellij.common.TextDocumentListenerAdapter;
 import com.microsoft.azure.toolkit.intellij.common.component.HighLightedCellRenderer;
-import com.microsoft.azure.toolkit.intellij.monitor.view.left.WorkspaceSelectionDialog;
 import com.microsoft.azure.toolkit.intellij.monitor.view.right.filter.KustoFilterComboBox;
 import com.microsoft.azure.toolkit.intellij.monitor.view.right.filter.TimeRangeComboBox;
 import com.microsoft.azure.toolkit.intellij.monitor.view.right.table.LogTable;
 import com.microsoft.azure.toolkit.intellij.monitor.view.right.table.LogTableModel;
 import com.microsoft.azure.toolkit.lib.Azure;
-import com.microsoft.azure.toolkit.lib.common.action.Action;
 import com.microsoft.azure.toolkit.lib.common.event.AzureEventBus;
 import com.microsoft.azure.toolkit.lib.common.exception.AzureToolkitRuntimeException;
 import com.microsoft.azure.toolkit.lib.common.messager.AzureMessager;
@@ -107,7 +104,7 @@ public class MonitorLogTablePanel {
 
     public void loadTableModel(@Nullable LogAnalyticsWorkspace selectedWorkspace, String queryString) {
         if (Objects.isNull(selectedWorkspace)) {
-            AzureMessager.getMessager().info(message("azure.monitor.info.selectWorkspace"), null, selectWorkspaceAction());
+            logTable.getEmptyText().setText(message("azure.monitor.info.selectWorkspaceTips"));
             return;
         }
         logTable.clearModel();
@@ -139,7 +136,6 @@ public class MonitorLogTablePanel {
 
     public void loadFilters(@Nullable LogAnalyticsWorkspace selectedWorkspace, String tableName) {
         if (Objects.isNull(selectedWorkspace)) {
-            AzureMessager.getMessager().info(message("azure.monitor.info.selectWorkspace"), null, selectWorkspaceAction());
             return;
         }
         timeRangePanel.setVisible(true);
@@ -291,18 +287,6 @@ public class MonitorLogTablePanel {
         } catch (final Exception e) {
             throw new AzureToolkitRuntimeException(e);
         }
-    }
-
-    private Action<?> selectWorkspaceAction() {
-        final Project project = ProjectManager.getInstance().getDefaultProject();
-        return new Action<>(Action.Id.of("user/monitor.select_workspace"))
-            .withLabel("Select")
-            .withHandler((s, e) -> AzureTaskManager.getInstance().runLater(() -> {
-                final WorkspaceSelectionDialog dialog = new WorkspaceSelectionDialog(project, null);
-                if (dialog.showAndGet()) {
-                    Optional.ofNullable(dialog.getWorkspace()).ifPresent(w -> AzureEventBus.emit("azure.monitor.change_workspace", w));
-                }
-            }));
     }
 
     // CHECKSTYLE IGNORE check FOR NEXT 1 LINES
